Fix event loops reported by Verilator UNOPTFLAT lints.
Decode is now split into a block which depends only on the instruction
bits, and a block which gates critical decode signals based on fetch
faults, invalidity etc.

Apply a similar transform to the gating of the uop counter update.

cxxrtl performance seems unchanged after removing the event loops, but
verilator and live-scheduled simulators should improve.