Implement fence.i fetch flush with an exception rather than a jump.
The issue with the jump implementation is that this can go out ahead of
a stalled load/store dphase. We allow fetches to go ahead in this case
to avoid passing load/store hready into fetch htrans (and it can help
perf too). This is problematic because in a multi-master AHB fabric the
address phase completion order does not necessarily match the data phase
completion. (This only happens in extremely contrived cases with a store
writing to a fetch nearby in the same basic block, but still a bug.)

Implementing refetch with a (kinda) exception guarantees ordering
against load/store data phases, which is what's actually required for
store-to-fetch ordering.

This mechanism will also allow more refetch triggers in future without
complicating jump/branch decode. For example, re-fetch of instructions
following writes to PMP config or the Zcmt table base CSR.